requirements.txt
llama-index
llama-index-embeddings-huggingface
python-dotenv
docx2txt
llama-index-llms-huggingface
gradio